refactoring import of `RankStatsWriter`
